refactor package structure (https://trello.com/c/eoTJyoV8)
