--- src/main/java/com/metalr2/model/followArtist/FollowedArtistsRepository.java
@@ -1,18 +1,0 @@
-package com.metalr2.model.followArtist;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface FollowedArtistsRepository extends JpaRepository<FollowedArtistEntity, Long> {
-
-  List<FollowedArtistEntity> findAllByPublicUserId(String publicUserId);
-
-  boolean existsByPublicUserIdAndArtistDiscogsId(String publicUserId, long artistDiscogsId);
-
-  Optional<FollowedArtistEntity> findByPublicUserIdAndArtistDiscogsId(String publicUserId, long artistDiscogsId);
-
-}
