+++ src/test/java/com/metalr2/web/controller/ArtistDetailsRestControllerIT.java
@@ -4,15 +4,12 @@ import com.metalr2.config.constants.Endpoints;
 import com.metalr2.model.user.UserEntity;
 import com.metalr2.model.user.UserRepository;
 import com.metalr2.model.user.UserRole;
+import com.metalr2.service.discogs.DiscogsArtistSearchRestClientImpl;
 import com.metalr2.web.DtoFactory;
 import com.metalr2.web.RestAssuredRequestHandler;
-import com.metalr2.web.controller.discogs.DiscogsArtistSearchRestClientImpl;
 import com.metalr2.web.dto.request.ArtistDetailsRequest;
-import com.metalr2.web.dto.request.ArtistSearchRequest;
-import com.metalr2.web.dto.response.ArtistNameSearchResponse;
-import com.metalr2.web.dto.response.ErrorResponse;
-import com.metalr2.web.dto.response.Pagination;
 import com.metalr2.web.dto.response.ArtistDetailsResponse;
+import com.metalr2.web.dto.response.ErrorResponse;
 import io.restassured.http.ContentType;
 import io.restassured.response.ValidatableResponse;
 import org.assertj.core.api.WithAssertions;
