+++ src/test/java/com/metalr2/web/controller/CustomErrorControllerIT.java
@@ -6,6 +6,7 @@ import com.metalr2.model.user.UserEntity;
 import com.metalr2.model.user.UserFactory;
 import com.metalr2.security.WebSecurity;
 import com.metalr2.service.user.UserService;
+import com.metalr2.web.controller.mvc.CustomErrorController;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
