+++ src/test/java/com/metalr2/web/controller/SearchArtistsControllerIT.java
@@ -5,6 +5,7 @@ import com.metalr2.config.constants.ViewNames;
 import com.metalr2.model.user.UserFactory;
 import com.metalr2.security.WebSecurity;
 import com.metalr2.service.user.UserService;
+import com.metalr2.web.controller.mvc.SearchArtistsController;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
